# Improve the implementation of mathics.core.list