* avoid building properties in the List constructor
* rewrite_apply_eval keeps untouched `self`, and returns a new object if the elements must be reevaluated.

Here there is a comparison between master and this branch for the combinatorial benchmark.

Pyston 2.3.4 (3.8.12)
NO_CYTHON=1
using SymPy 1.8, mpmath 1.2.1, numpy 1.21.4


|   test                                        |   master           |   this branch            |
| --------------------------------------------- | ------------------ |------------------ |
| load_combinatorica 				|	207.29 ms    |  181.15 ms        |   
| test_permutations_1_1   		       	|	881.47 ms    | 	930.90 ms   	 |
| test_permutations_groups_1_2   		|	7593.88 ms   | 	5642.51 ms  	 |
| test_inversions_and_inversion_vectors_1_3   	|	713.37 ms    | 	530.34 ms   	 |
| test_special_classes_of_permutations_1_4   	|	404.48 ms    | 	291.73 ms   	 |
| test_combinations_1_5   			|	634.20 ms    | 	456.43 ms   	 |
| test_2_1_to_2_3   				|	93.29 ms     | 	67.99 ms    	 |
| test_combinatorica_rest   			|	129.06 ms    | 	124.77 ms   	 |